requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.8
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.24.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
mcp==1.3.0
multidict==6.6.3
openai==0.28.1
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sse-starlette==2.4.1
starlette==0.45.3
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.20.1